Move ITerminalInstance.runRecent into its own file
Fixes #159618
